fix(ci): mesh-lab worker verify and testing-strategy gate
- Worker tier checks /api/status instead of fleet routes removed from open API
- Fold HttpRemoteBrickCatalog coverage into existing capability fallback tests
- Rename KnowledgeQueryService tests to avoid new GapCoverage file gate
