+++ docs/bricks/unknown.md
@@ -2,18 +2,18 @@
 
 ## Behavior
 
-Adapted from promotion b16850fbd8f54c73bd1ae37ce6d3acbc.
+Adapted from promotion 02b2c57f73ed45f688f1928684f47dd5.
 
-**Last updated:** 2026-05-28
+**Last updated:** 2026-06-08
 
 ## Changelog
 
 ```markdown
 # Changelog
 
-## 2026-05-27 – 2026-05-28
+## 2026-06-07 – 2026-06-08
 
-- **unknown**: Fixed EmptyCatch in EmptyCatch.cs (2026-05-28 02:54)
+- **unknown**: Fixed EmptyCatch in EmptyCatch.cs (2026-06-08 15:08)
 
 
 ```
